Added HTTP requests to the unit tests
Added basic unit tests that send HTTP requests to the server.
These could be extended to allow testing stage after a deployment.
We could also use tests like this to validate the security settings of production

These tests use the REST API as the entry point to exersise code paths that
are not covered by testing of the isolated code.

These test can also catch cases where the server fails to start.

Also fixed the jshint for the test code so we validate that code to.

Change-Id: Icca9319d16cbd5e94a99082f8a64ed6b44c2d624